boiler colors now change properly
