+++ src/pages/books/the-alchemy-of-air/PreheatsReagents.tsx
@@ -3,6 +3,8 @@ import { Particle } from "@/pages/books/the-alchemy-of-air/Particle";
 import React from "react";
 import { startAnimationEventName } from "@/common/animations";
 import { Block } from "@/pages/books/the-alchemy-of-air/Block";
+import { Queue } from "@/common/queue";
+import { generateGradient } from "@/utils/generateGradient";
 
 const initialParticles = [
     // new Particle(100, 100, 140, 3, 3, 0, 0, 5, "red", 50),
@@ -22,6 +24,12 @@ export default function PreheatsReagents() {
     const particles = React.useRef<Particle[]>([]);
     const boilerRef = React.useRef<HTMLDivElement>(null);
     const [startAnimation, setStartAnimation] = React.useState<boolean>(false);
+    const colorArray = React.useRef<Array<string>>([]);
+
+    React.useEffect(() => {
+        colorArray.current = generateGradient(3);
+        colorArray.current.reverse(); // after the rversal, red is on the left, and blue is on the right
+    }, []);
 
     // ref to the ParticleSimulationCanvas
     React.useEffect(() => {
@@ -31,9 +39,19 @@ export default function PreheatsReagents() {
         const boiler = boilerRef.current;
         const spawnParticle = () => {
             const particleVy = Math.floor(Math.random() * 5) + 1;
-            particles.current.push(new Particle(30, 30, 40, 3, particleVy, 0, 0, 5, "red", 200));
+            particles.current.push(new Particle(30, 30, 40, 3, particleVy, 0, 0, 5, "blue", 200));
 
-            const spawnDelay = Math.floor(Math.random() * 201) + 200; // randomize the spawn delay
+            // loop through the spawned particles, and set them to 
+            const numParticles =  particles.current.length
+            const numColors = colorArray.current.length;
+            // doing it this way makes it so the particles that are just spawned are blue, and the later ones are red
+            for(let i = 0; i <numParticles; i++) {
+                const ithColorToPick = Math.max(i - (numParticles - numColors), 0);
+                particles.current[i].color = colorArray.current[ithColorToPick];
+            }
+            // console.log(particles.current)
+
+            const spawnDelay = Math.floor(Math.random() * 401) + 200; // randomize the spawn delay
             timeoutId.current = setTimeout(spawnParticle, spawnDelay);
         };
